requirements.txt
fastapi[all]
python-dotenv
openai
supabase
requests
uvicorn
pydantic
